fix: configure assembly task concurrent restriction
This to prevent the task to fail due to resource starvation.
For instance cpu and to many open files. The issue will most
likely occur on multi-module project where we do assembly on
more than one project. Setting the default to one will ensure
that it will work and can be overridden if needed.